+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/AICommitAction.kt
@@ -1,63 +1,36 @@
 package com.github.blarc.ai.commits.intellij.plugin
 
-import com.github.blarc.ai.commits.intellij.plugin.AICommitsBundle.message
-import com.github.blarc.ai.commits.intellij.plugin.notifications.Notification
-import com.github.blarc.ai.commits.intellij.plugin.notifications.sendNotification
-import com.github.blarc.ai.commits.intellij.plugin.settings.ProjectSettings
-import com.intellij.openapi.actionSystem.ActionUpdateThread
-import com.intellij.openapi.actionSystem.AnAction
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.components.service
+import com.github.blarc.ai.commits.intellij.plugin.settings.AppSettings2
+import com.github.blarc.ai.commits.intellij.plugin.settings.clients.LLMClientConfiguration
+import com.intellij.openapi.actionSystem.*
 import com.intellij.openapi.project.DumbAware
-import com.intellij.openapi.vcs.VcsDataKeys
-import com.intellij.openapi.vcs.ui.CommitMessage
-import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
 
-class AICommitAction : AnAction(), DumbAware {
+class AICommitAction : SplitButtonAction(object : ActionGroup() {
+    override fun getChildren(e: AnActionEvent?): Array<AnAction> {
+        val configurations = AppSettings2.instance.llmClientConfigurations.sortedWith(
+            compareBy<LLMClientConfiguration> {
+                it.id != AppSettings2.instance.activeLlmClientId
+            }.thenBy {
+                it.name
+            }
+        )
 
+        val actions = mutableListOf<AnAction>()
+        if (configurations.isNotEmpty()) {
+            actions.add(configurations.first())
+
+            if (configurations.size > 1) {
+                actions.add(Separator.getInstance())
+            }
+
+            actions.addAll(configurations.drop(1))
+        }
+
+        return actions.toTypedArray()
+    }
+}), DumbAware {
 
     override fun getActionUpdateThread(): ActionUpdateThread {
         return ActionUpdateThread.EDT
-    }
-
-    override fun update(e: AnActionEvent) {
-        e.project?.service<ProjectSettings>()?.getActiveLLMClientConfiguration()?.let {
-            if (it.getGenerateCommitMessageJob()?.isActive == true) {
-                e.presentation.icon = Icons.Process.STOP.getThemeBasedIcon()
-            } else {
-                e.presentation.icon = it.getClientIcon()
-                e.presentation.text = message("action.tooltip", it.name)
-            }
-        }
-    }
-
-    override fun actionPerformed(e: AnActionEvent) {
-        val project = e.project ?: return
-
-        val llmClient = project.service<ProjectSettings>().getActiveLLMClientConfiguration()
-        if (llmClient == null) {
-            sendNotification(Notification.clientNotSet())
-            return
-        }
-
-        val generateCommitMessageJob = llmClient.getGenerateCommitMessageJob()
-        if (generateCommitMessageJob?.isActive == true) {
-            generateCommitMessageJob.cancel()
-            return
-        }
-
-        val commitWorkflowHandler = e.getData(VcsDataKeys.COMMIT_WORKFLOW_HANDLER) as AbstractCommitWorkflowHandler<*, *>?
-        if (commitWorkflowHandler == null) {
-            sendNotification(Notification.noCommitMessage())
-            return
-        }
-
-        val commitMessage = VcsDataKeys.COMMIT_MESSAGE_CONTROL.getData(e.dataContext) as CommitMessage?
-        if (commitMessage == null) {
-            sendNotification(Notification.noCommitMessage())
-            return
-        }
-
-        llmClient.generateCommitMessage(commitWorkflowHandler, commitMessage, project)
     }
 }