+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added 
+
+- Generate commit action button is now a split button with a dropdown from which the user can change the LLM client from the commit dialog.
+
 ## [2.11.1] - 2025-04-25
 
 ### Fixed